% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d3tree3.R
\name{d3tree3Output}
\alias{d3tree3Output}
\title{Widget output function for use in Shiny}
\usage{
d3tree3Output(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

