% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3tree4.R
\name{d3tree4Output}
\alias{d3tree4Output}
\title{Widget output function for use in Shiny}
\usage{
d3tree4Output(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}
