% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/d3tree.R
\name{d3treeOutput}
\alias{d3treeOutput}
\title{Widget output function for use in Shiny}
\usage{
d3treeOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

