% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/d3tree.R
\name{renderD3tree}
\alias{renderD3tree}
\title{Widget render function for use in Shiny}
\usage{
renderD3tree(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

