% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d3tree3.R
\name{renderD3tree3}
\alias{renderD3tree3}
\title{Widget render function for use in Shiny}
\usage{
renderD3tree3(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

