% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3tree4.R
\name{renderd3tree4}
\alias{renderd3tree4}
\title{Widget render function for use in Shiny}
\usage{
renderd3tree4(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}
